refactor: 연결한 도메인 객체에 따라 타입, 데이터 구조 수정
